Short-Circuit Choose and Fix typos
When `monsql.drop_table` is invoked, internally we first evaluate
`silent`. If it is true, we continue to evaluate `.is_table_existed`.
If false, we ignore the second evaluation.This will avoid accessing
the database (in evaluating `.is_table_existed`) everytime
`monsql.drop_table` is invoked.

For more information on short-circuit operators, see
https://docs.python.org/2/library/stdtypes.html#boolean-operations-and-or-not

Some minor typo errors are fixed. `silent` was mispelled as `slient`.